+++ src/BaGet.Web/Extensions/IServiceCollectionExtensions.cs
@@ -1,12 +1,9 @@
 using System;
-using System.IO;
 using BaGet.Core;
 using BaGet.Web;
 using Microsoft.AspNetCore.Mvc;
 using Microsoft.AspNetCore.Mvc.Razor.RuntimeCompilation;
 using Microsoft.Extensions.DependencyInjection;
-using Microsoft.Extensions.FileProviders;
-using Microsoft.Extensions.Hosting;
 using Microsoft.Extensions.Options;
 
 namespace BaGet
@@ -40,29 +37,6 @@ namespace BaGet
             services.AddBaGetApplication(configureAction);
 
             return services;
-        }
-
-        private class ConfigureRazorRuntimeCompilation : IConfigureOptions<MvcRazorRuntimeCompilationOptions>
-        {
-            private readonly IHostEnvironment _env;
-
-            public ConfigureRazorRuntimeCompilation(IHostEnvironment env)
-            {
-                _env = env ?? throw new ArgumentNullException(nameof(env));
-            }
-
-            public void Configure(MvcRazorRuntimeCompilationOptions options)
-            {
-                var path = Path.Combine(_env.ContentRootPath, "..", "BaGet.Web");
-
-                // Try to enable Razor "hot reload".
-                if (!_env.IsDevelopment()) return;
-                if (!Directory.Exists(path)) return;
-
-                var provider = new PhysicalFileProvider(Path.GetFullPath(path));
-
-                options.FileProviders.Add(provider);
-            }
         }
     }
 }
